Refactor the MavenResolverClient class
