+++ misc/maven-resolver/src/main/java/org/ballerinalang/maven/bala/client/model/ToolMavenMetadata.java
@@ -1,0 +1,89 @@
+/*
+ * Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.ballerinalang.maven.bala.client.model;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Data class representing tool metadata parsed from a Maven maven-metadata.xml file.
+ *
+ * @since 2201.8.0
+ */
+public class ToolMavenMetadata {
+    private String groupId;
+    private String artifactId;
+    private String org;
+    private String name;
+    private List<ToolVersion> versions;
+
+    public ToolMavenMetadata() {
+        this.versions = new ArrayList<>();
+    }
+
+    public String getGroupId() {
+        return groupId;
+    }
+
+    public void setGroupId(String groupId) {
+        this.groupId = groupId;
+    }
+
+    public String getArtifactId() {
+        return artifactId;
+    }
+
+    public void setArtifactId(String artifactId) {
+        this.artifactId = artifactId;
+    }
+
+    public String getOrg() {
+        return org;
+    }
+
+    public void setOrg(String org) {
+        this.org = org;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public List<ToolVersion> getVersions() {
+        return versions;
+    }
+
+    public void setVersions(List<ToolVersion> versions) {
+        this.versions = versions;
+    }
+
+    @Override
+    public String toString() {
+        return "ToolMavenMetadata{" +
+                "groupId='" + groupId + '\'' +
+                ", artifactId='" + artifactId + '\'' +
+                ", org='" + org + '\'' +
+                ", name='" + name + '\'' +
+                ", versions=" + versions +
+                '}';
+    }
+}
